requirements.txt
Django==1.8
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
MySQL-python==1.2.5
PyJWT==1.5.2
